Kill zombies (#127)
Running `testarmada-nightwatch-extra@5.1.1` with Appium, I'm running into the nightwatch process hanging open. It seems to be hit or miss but happens more often than not. It will pass the test, but never complete.

I saw https://github.com/TestArmada/nightwatch-extra/pull/94 had been added in `4.x` to solve this, but was removed in `5.x`. Any reason?

Adding it back in (**this is a copy+paste of what was previously in there**) solves the issue:
![image](https://user-images.githubusercontent.com/916905/51124528-58fc9f80-17d3-11e9-99a5-1337aa9effdf.png)

Happy to create a demo project to recreate this and refactor this out if you think it's needed.